+++ src/interfaces/ILiquidityPool.sol
@@ -75,7 +75,7 @@ interface ILiquidityPool {
     function payProtocolFees(uint128 _protocolFees) external;
     function addEthAmountLockedForWithdrawal(uint128 _amount) external;
     function transferLockedEthForPriority(uint128 _amount) external;
-    function returnLockedEth(uint256 _amount) external payable;
+    function returnLockedEth(uint128 _amount) external payable;
 
     function pauseContract() external;
     function burnEEthShares(uint256 shares) external;
